pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "schema-drift"
version = "2.0.0"
description = "Track why your schema changed, not just what changed."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9"
keywords = ["database", "schema", "migrations", "diff", "postgres", "sqlite"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Database",
    "Topic :: Software Development :: Libraries",
]
dependencies = []

[project.optional-dependencies]
postgres = ["psycopg2-binary>=2.9"]
mysql = ["mysql-connector-python>=8.0"]
oracle = ["oracledb>=1.0"]
dev = ["pytest", "pytest-cov", "ruff"]

[project.scripts]
schema-drift = "schema_drift.cli:main"

[project.urls]
Homepage = "https://github.com/woxlo1/schema-drift"
Repository = "https://github.com/woxlo1/schema-drift"

[tool.ruff]
line-length = 100
